backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.0
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
email-validator==2.2.0
